+++ src/api/objects/toot.ts
@@ -921,9 +921,15 @@ export default class Toot implements TootObj {
         return deduped;
     }
 
+    // Get rid of toots we never want to see again
     static async removeInvalidToots(toots: Toot[], logger: Logger): Promise<Toot[]> {
         const serverSideFilters = (await MastoApi.instance.getServerSideFilters()) || [];
         return filterWithLog(toots, t => t.isValidForFeed(serverSideFilters), logger, 'invalid', 'Toot');
+    }
+
+    // Filter an array of toots down to just the retoots
+    static onlyRetoots(toots: Toot[]): Toot[] {
+        return toots.filter(toot => toot.reblog);
     }
 
     // Return a new array of a toot property collected and uniquified from an array of toots