+++ src/enums.ts
@@ -84,9 +84,3 @@ export enum TrendingType {
     STATUSES = "statuses",
     TAGS = "tags"
 };
-
-// Both filter option property names as well as demo app gradient config keys
-export const FILTER_OPTION_DATA_SOURCES = [
-    ...Object.values(TagTootsCacheKey),  // TODO: these are really the wrong cache keys for the use case but it's consistent w/demo app for now
-    ScoreName.FAVOURITED_ACCOUNTS,
-] as const;